+++ flutter_inappwebview_platform_interface/lib/src/types/proxy_relay_hop.g.dart
@@ -18,21 +18,18 @@ class ProxyRelayHop {
   ///
   ///**Officially Supported Platforms/Implementations**:
   ///- iOS WKWebView
-  ///- macOS WKWebView
   Map<String, String>? additionalHTTPHeaders;
 
   ///A URL or host endpoint identifying the relay server accessible using HTTP/2.
   ///
   ///**Officially Supported Platforms/Implementations**:
   ///- iOS WKWebView
-  ///- macOS WKWebView
   String? http2RelayEndpoint;
 
   ///A URL or host endpoint identifying the relay server accessible using HTTP/3.
   ///
   ///**Officially Supported Platforms/Implementations**:
   ///- iOS WKWebView
-  ///- macOS WKWebView
   String? http3RelayEndpoint;
   ProxyRelayHop({
     this.http3RelayEndpoint,